requirements.txt
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
et-xmlfile==1.1.0
google-api-core==2.14.0
google-api-python-client==2.109.0
google-auth==2.25.1
google-auth-httplib2==0.1.1
google-auth-oauthlib==1.1.0
googleapis-common-protos==1.61.0
gspread==5.12.2
httplib2==0.22.0
idna==3.6
numpy==1.26.2
oauthlib==3.2.2
openpyxl==3.1.2
pandas==2.1.3
protobuf==4.25.1
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
rapidfuzz==3.5.2
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
six==1.16.0
thefuzz==0.20.0
tqdm==4.66.1
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
